fix: fix file directory
