+++ src/services/project/project-detail/ProjectDetailPage.vue
@@ -147,8 +147,8 @@ import { ALERT_STATE } from '@/services/alert-manager/lib/config';
 import MaintenanceHappeningList from '@/services/project/project-detail/modules/MaintenanceHappeningList.vue';
 import MaintenanceWindowFormModal from '@/services/project/project-detail/modules/MaintenanceWindowFormModal.vue';
 import ProjectFormModal from '@/services/project/project-detail/modules/ProjectFormModal.vue';
-import { useProjectDetailPageStore } from '@/services/project/project-detail/store/project-detail-page-store';
 import { PROJECT_ROUTE } from '@/services/project/route-config';
+import { useProjectDetailPageStore } from '@/services/project/store/project-detail-page-store';
 import type { ProjectModel } from '@/services/project/type';
 
 export default {
